+++ app.py
@@ -6,15 +6,14 @@ from flask import url_for
 from authlib.integrations.flask_client import OAuth
 from flask_wtf.csrf import CSRFProtect
 from urllib.parse import urlencode  # Abweichung von OAuth-Quickstarts
-from functools import wraps
 import auth
 import json
 import sqlite3 as sql
+import decorators
 
 app = Flask(__name__)
 csrf = CSRFProtect()
 csrf.init_app(app)
-
 
 # Secret key
 app.config['SECRET_KEY'] = "T5BPYMJD9GVKURSGTAXC"
@@ -33,126 +32,6 @@ auth0 = oauth.register(
     },
 )
 
-# put your user_id here
-ADMINS = ["github|59766382"]
-
-
-# Decorator
-def requires_auth(f):
-    @wraps(f)
-    def decorated(*args, **kwargs):
-        if 'profile' not in session:
-            return redirect(url_for("error"))
-        return f(*args, **kwargs)
-
-    return decorated
-
-
-# Helper Functions
-# Checks if user_id from session is present in userstore from db and member of specific role
-def is_admin(f):
-    con = sql.connect('database.db')
-    con.row_factory = sql.Row
-    cur = con.cursor()
-    userstore = cur.execute("SELECT * FROM users;").fetchall()
-    for i in userstore:
-        if session["profile"]["user_id"] in i and session["profile"][
-                "user_id"] in ADMINS:
-            return True
-        else:
-            return False
-
-
-def is_dozent(f):
-    con = sql.connect('database.db')
-    con.row_factory = sql.Row
-    cur = con.cursor()
-    userstore = cur.execute("SELECT * FROM users;").fetchall()
-    for i in userstore:
-        if session["profile"]["user_id"] in i and "is_dozent" in i:
-            return True
-        else:
-            return False
-
-
-def is_student(f):
-    con = sql.connect('database.db')
-    con.row_factory = sql.Row
-    cur = con.cursor()
-    userstore = cur.execute("SELECT * FROM users;").fetchall()
-    for i in userstore:
-        if session["profile"]["user_id"] in i and "is_student" in i:
-            return True
-        else:
-            return False
-
-
-# Authorization decorators
-def admin_only(f):
-    @wraps(f)
-    def decorated_function(*args, **kwargs):
-
-        if not is_admin(f):
-            return redirect(url_for("error"))
-        return f(*args, **kwargs)
-
-    return decorated_function
-
-
-def not_admin_only(f):
-    @wraps(f)
-    def decorated_function(*args, **kwargs):
-
-        if is_admin(f):
-            return redirect(url_for("error"))
-        return f(*args, **kwargs)
-
-    return decorated_function
-
-
-def dozent_only(f):
-    @wraps(f)
-    def decorated_function(*args, **kwargs):
-
-        if not is_dozent(f):
-            return redirect(url_for("error"))
-        return f(*args, **kwargs)
-
-    return decorated_function
-
-
-def not_dozent_only(f):
-    @wraps(f)
-    def decorated_function(*args, **kwargs):
-
-        if is_dozent(f):
-            return redirect(url_for("error"))
-        return f(*args, **kwargs)
-
-    return decorated_function
-
-
-def student_only(f):
-    @wraps(f)
-    def decorated_function(*args, **kwargs):
-
-        if not is_student(f):
-            return redirect(url_for("error"))
-        return f(*args, **kwargs)
-
-    return decorated_function
-
-
-def not_student_only(f):
-    @wraps(f)
-    def decorated_function(*args, **kwargs):
-
-        if is_student(f):
-            return redirect(url_for("error"))
-        return f(*args, **kwargs)
-
-    return decorated_function
-
 
 # Routes
 @app.route('/', methods=['GET'])
@@ -161,7 +40,7 @@ def index():
 
 
 @app.route('/home', methods=['GET'])
-@requires_auth
+@decorators.requires_auth
 def home():
     return render_template('home.html',
                            userinfo=session['profile'],
@@ -170,8 +49,8 @@ def home():
 
 
 @app.route('/single', methods=['GET'])
-@requires_auth
-@student_only
+@decorators.requires_auth
+@decorators.student_only
 def single():
     return render_template('single.html',
                            userinfo=session['profile'],
@@ -180,8 +59,8 @@ def single():
 
 
 @app.route('/multi', methods=['GET'])
-@requires_auth
-@student_only
+@decorators.requires_auth
+@decorators.student_only
 def multi():
     return render_template('multi.html',
                            userinfo=session['profile'],
@@ -190,7 +69,7 @@ def multi():
 
 
 @app.route('/dashboard', methods=['GET'])
-@requires_auth
+@decorators.requires_auth
 def dashboard():
     return render_template('dashboard.html',
                            userinfo=session['profile'],
@@ -199,8 +78,8 @@ def dashboard():
 
 
 @app.route('/rank', methods=['GET'])
-@requires_auth
-@student_only
+@decorators.requires_auth
+@decorators.student_only
 def rank():
     return render_template('rank.html',
                            userinfo=session['profile'],
@@ -209,8 +88,8 @@ def rank():
 
 
 @app.route('/list', methods=['GET'])
-@requires_auth
-@admin_only
+@decorators.requires_auth
+@decorators.admin_only
 def list():
     # link sql database
     con = sql.connect("database.db")
@@ -226,7 +105,7 @@ def list():
 
 
 @app.route('/about', methods=['GET'])
-@requires_auth
+@decorators.requires_auth
 def about():
     return render_template('about.html',
                            userinfo=session['profile'],
@@ -267,10 +146,11 @@ def callback_handling():
 
     with sql.connect("database.db") as con:
         cur = con.cursor()
-        cur.execute("INSERT OR IGNORE INTO users (id,username,role) VALUES (?,?,?)",
-                    (user_id, username, role))
+        cur.execute(
+            "INSERT OR IGNORE INTO users (id,username,role) VALUES (?,?,?)",
+            (user_id, username, role))
         con.commit()
-        
+
         return redirect('/dashboard')
